refactor: fold the adapter checks into `memo release check`
The first cut shipped a standalone `scripts/adapter_matrix.py` with its own
version-parity check. That check was both redundant and weaker than what
already existed: `release_check_report` validates version parity across the
plugin manifests, server.json (including extra packages), the mcpb manifests
and archive, install pins, the Homebrew formula and the CHANGELOG. Two gates
that can disagree about the same fact is worse than one — dropped it.

What is genuinely uncovered is the drift that carries no version and fails
silently, so the three remaining checks move into `src/memo/adapter_matrix.py`
and run as part of `memo release check`. The standalone script had no
independence to trade away for this: `hook-commands-resolve` resolves against
the live Click registry, so it needed the package importable either way.

Also fixes an over-strict reading that the release tests caught: an absent
surface is not drift. A missing hook graph means there are no commands to
resolve, and a repo with no `.mcp.json` has no dims to compare — both are
vacuously clean. `skipped` (which adapter_issues reports as an issue, never a
silent pass) is now reserved for a surface that exists but could not be
verified.

Verified end to end: breaking a hook command in hooks/hooks.json makes the real
`memo release check` fail with the offending subcommand named.
