+++ tests/test_adapter_matrix.py
@@ -2,38 +2,21 @@
 
 Each test mutates exactly one surface of a synthetic repo and asserts the
 matching check flips to fail — a gate that only ever passes proves nothing.
+
+Version parity is deliberately NOT tested here: `memo release check` owns it
+(`cli_release._VERSION_TARGETS` + `_check_changelog` + `_check_mcpb_*` +
+`_check_formula`), and a second, weaker opinion that can disagree with the real
+gate is worse than no opinion.
 """
 
 from __future__ import annotations
 
-import importlib.util
 import json
-import sys
 from pathlib import Path
 
 import pytest
 
-ROOT = Path(__file__).resolve().parent.parent
-
-
-def _load_adapter_matrix():
-    """Load scripts/adapter_matrix.py the way tests/test_dev_audit.py loads its
-    script: by path, since scripts/ is not an installed package."""
-    script = ROOT / "scripts" / "adapter_matrix.py"
-    assert script.is_file(), "scripts/adapter_matrix.py must exist"
-    spec = importlib.util.spec_from_file_location("memo_adapter_matrix", script)
-    assert spec and spec.loader
-    module = importlib.util.module_from_spec(spec)
-    sys.modules[spec.name] = module
-    spec.loader.exec_module(module)
-    return module
-
-
-_matrix = _load_adapter_matrix()
-main = _matrix.main
-run = _matrix.run
-
-VERSION = "1.2.3"
+from memo.adapter_matrix import adapter_issues, run
 
 
 def _write_json(path: Path, payload: object) -> None:
@@ -43,31 +26,19 @@ def _write_json(path: Path, payload: object) -> None:
 
 @pytest.fixture
 def repo(tmp_path: Path) -> Path:
-    """A minimal repo where every adapter-matrix check passes."""
-    (tmp_path / "pyproject.toml").write_text(
-        f'[project]\nname = "mlx-memo"\nversion = "{VERSION}"\n\n[tool.ruff]\nversion = "ignored"\n',
-        encoding="utf-8",
-    )
-    (tmp_path / "CHANGELOG.md").write_text(
-        f"# Changelog\n\n## [Unreleased]\n\n## [{VERSION}] - 2026-01-01\n",
-        encoding="utf-8",
-    )
-    _write_json(tmp_path / ".claude-plugin" / "plugin.json", {"name": "memo", "version": VERSION})
+    """A minimal repo where every adapter check passes."""
     _write_json(
         tmp_path / ".claude-plugin" / "marketplace.json",
         {"name": "memo", "plugins": [{"name": "memo", "source": "./"}]},
     )
+    _write_json(tmp_path / ".claude-plugin" / "plugin.json", {"name": "memo"})
     _write_json(
         tmp_path / "plugins" / "memo" / ".codex-plugin" / "plugin.json",
-        {"name": "memo", "version": VERSION, "mcpServers": "./.mcp.json"},
+        {"name": "memo", "mcpServers": "./.mcp.json"},
     )
     _write_json(
         tmp_path / "plugins" / "memo" / ".mcp.json",
         {"mcpServers": {"memo": {"command": "memo-mcp", "env": {"MEMO_NONINTERACTIVE": "1"}}}},
-    )
-    _write_json(
-        tmp_path / "server.json",
-        {"name": "io.github.jagoff/memo", "version": VERSION, "packages": [{"version": VERSION}]},
     )
     _write_json(
         tmp_path / ".mcp.json",
@@ -119,57 +90,10 @@ def _status(root: Path, check_id: str) -> str:
 
 def test_clean_repo_passes_every_check(repo: Path) -> None:
     assert [c.check_id for c in run(repo) if not c.ok] == []
-    assert main(["--check", "--root", str(repo)]) == 0
+    assert adapter_issues(repo) == []
 
 
-def test_version_drift_in_one_manifest_fails(repo: Path) -> None:
-    _write_json(repo / "server.json", {"version": "9.9.9", "packages": [{"version": VERSION}]})
-
-    assert _status(repo, "version-parity") == "fail"
-    assert main(["--check", "--root", str(repo)]) == 1
-
-
-def test_changelog_missing_the_released_version_fails(repo: Path) -> None:
-    (repo / "CHANGELOG.md").write_text("# Changelog\n\n## [Unreleased]\n", encoding="utf-8")
-
-    assert _status(repo, "version-parity") == "fail"
-
-
-def test_embedder_dims_not_matching_model_size_fails(repo: Path) -> None:
-    # 4B is 2560-dim; 1024 belongs to the 0.6B model. This exact mismatch
-    # corrupts the vec0 table (MLX invariant 3).
-    _write_json(
-        repo / ".mcp.json",
-        {
-            "mcpServers": {
-                "memo": {
-                    "env": {
-                        "MEMO_EMBEDDER_MODEL": "mlx-community/Qwen3-Embedding-4B-4bit-DWQ",
-                        "MEMO_EMBEDDER_DIMS": "1024",
-                    }
-                }
-            }
-        },
-    )
-
-    assert _status(repo, "embedder-dims-parity") == "fail"
-
-
-def test_pinned_model_without_dims_fails(repo: Path) -> None:
-    _write_json(
-        repo / ".mcp.json",
-        {"mcpServers": {"memo": {"env": {"MEMO_EMBEDDER_MODEL": "Qwen3-Embedding-8B"}}}},
-    )
-
-    assert _status(repo, "embedder-dims-parity") == "fail"
-
-
-def test_config_pinning_neither_model_nor_dims_passes(repo: Path) -> None:
-    # The shipped plugin config deliberately pins no model: the installed index
-    # is self-describing and adopts its own profile. That must stay legal.
-    _write_json(repo / ".mcp.json", {"mcpServers": {"memo": {"env": {"MEMO_NONINTERACTIVE": "1"}}}})
-
-    assert _status(repo, "embedder-dims-parity") == "pass"
+# --- hooks ↔ CLI --------------------------------------------------------------
 
 
 def test_hook_firing_an_unregistered_subcommand_fails(repo: Path) -> None:
@@ -189,6 +113,7 @@ def test_hook_firing_an_unregistered_subcommand_fails(repo: Path) -> None:
     )
 
     assert _status(repo, "hook-commands-resolve") == "fail"
+    assert any("onse" in issue for issue in adapter_issues(repo))
 
 
 def test_hook_firing_an_unregistered_top_level_command_fails(repo: Path) -> None:
@@ -229,6 +154,95 @@ def test_hook_env_prefix_and_flags_do_not_confuse_resolution(repo: Path) -> None
     assert _status(repo, "hook-commands-resolve") == "pass"
 
 
+def test_a_missing_hook_graph_is_not_drift(repo: Path) -> None:
+    """No hook graph means no commands to resolve. Treating absence as drift
+    made every partial tree (the release tests' synthetic repos, a sparse
+    checkout) fail the gate for something it does not measure."""
+    (repo / "hooks" / "hooks.json").unlink()
+
+    assert _status(repo, "hook-commands-resolve") == "pass"
+    assert adapter_issues(repo) == []
+
+
+# --- embedder dims (MLX invariant 3) ------------------------------------------
+
+
+def test_embedder_dims_not_matching_model_size_fails(repo: Path) -> None:
+    # 4B is 2560-dim; 1024 belongs to the 0.6B model. This exact mismatch
+    # corrupts the vec0 table.
+    _write_json(
+        repo / ".mcp.json",
+        {
+            "mcpServers": {
+                "memo": {
+                    "env": {
+                        "MEMO_EMBEDDER_MODEL": "mlx-community/Qwen3-Embedding-4B-4bit-DWQ",
+                        "MEMO_EMBEDDER_DIMS": "1024",
+                    }
+                }
+            }
+        },
+    )
+
+    assert _status(repo, "embedder-dims-parity") == "fail"
+
+
+def test_pinned_model_without_dims_fails(repo: Path) -> None:
+    _write_json(
+        repo / ".mcp.json",
+        {"mcpServers": {"memo": {"env": {"MEMO_EMBEDDER_MODEL": "Qwen3-Embedding-8B"}}}},
+    )
+
+    assert _status(repo, "embedder-dims-parity") == "fail"
+
+
+def test_dims_pinned_without_a_model_fails(repo: Path) -> None:
+    _write_json(
+        repo / ".mcp.json", {"mcpServers": {"memo": {"env": {"MEMO_EMBEDDER_DIMS": "2560"}}}}
+    )
+
+    assert _status(repo, "embedder-dims-parity") == "fail"
+
+
+def test_config_pinning_neither_model_nor_dims_passes(repo: Path) -> None:
+    # The shipped plugin config deliberately pins no model: the installed index
+    # is self-describing and adopts its own profile. That must stay legal.
+    _write_json(repo / ".mcp.json", {"mcpServers": {"memo": {"env": {"MEMO_NONINTERACTIVE": "1"}}}})
+
+    assert _status(repo, "embedder-dims-parity") == "pass"
+
+
+def test_a_repo_with_no_mcp_config_is_not_drift(repo: Path) -> None:
+    """`skipped` is reserved for a surface that exists but could not be
+    verified, because adapter_issues reports it. An absent surface is nothing
+    to check — conflating the two made every partial tree fail the gate."""
+    (repo / ".mcp.json").unlink()
+    (repo / "plugins" / "memo" / ".mcp.json").unlink()
+
+    assert _status(repo, "embedder-dims-parity") == "pass"
+    assert [i for i in adapter_issues(repo) if "embedder-dims" in i] == []
+
+
+def test_every_known_model_size_maps_to_its_own_dims(repo: Path) -> None:
+    for model, dims in (
+        ("Qwen3-Embedding-0.6B", 1024),
+        ("Qwen3-Embedding-4B", 2560),
+        ("Qwen3-Embedding-8B", 4096),
+    ):
+        _write_json(
+            repo / ".mcp.json",
+            {
+                "mcpServers": {
+                    "memo": {"env": {"MEMO_EMBEDDER_MODEL": model, "MEMO_EMBEDDER_DIMS": str(dims)}}
+                }
+            },
+        )
+        assert _status(repo, "embedder-dims-parity") == "pass", model
+
+
+# --- referenced paths ---------------------------------------------------------
+
+
 def test_codex_manifest_pointing_at_a_missing_mcp_config_fails(repo: Path) -> None:
     (repo / "plugins" / "memo" / ".mcp.json").unlink()
 
@@ -244,5 +258,40 @@ def test_marketplace_source_without_a_plugin_manifest_fails(repo: Path) -> None:
     assert _status(repo, "referenced-paths-exist") == "fail"
 
 
-def test_missing_repo_root_is_a_usage_error(tmp_path: Path) -> None:
-    assert main(["--check", "--root", str(tmp_path / "nope")]) == 2
+# --- integration with the real gate -------------------------------------------
+
+
+def test_release_check_surfaces_adapter_drift(repo: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+    """`memo release check` must actually run these checks — a module nobody
+    calls is the failure mode this whole change exists to avoid."""
+    from memo import cli_release
+
+    _write_json(
+        repo / "hooks" / "hooks.json",
+        {
+            "hooks": {
+                "Stop": [{"hooks": [{"type": "command", "command": "memo definitely-not-a-cmd"}]}]
+            }
+        },
+    )
+    # Short-circuit the version half; this test is about the adapter half.
+    monkeypatch.setattr(cli_release, "_read_current_version", lambda _repo: "1.2.3")
+    monkeypatch.setattr(cli_release, "_check_json_version_targets", lambda *a, **k: None)
+    monkeypatch.setattr(cli_release, "_check_install_pins", lambda *a, **k: None)
+    monkeypatch.setattr(cli_release, "_check_additional_server_packages", lambda *a, **k: None)
+    monkeypatch.setattr(cli_release, "_check_mcpb_manifest", lambda *a, **k: None)
+    monkeypatch.setattr(cli_release, "_check_mcpb_node_manifest", lambda *a, **k: None)
+    monkeypatch.setattr(cli_release, "_check_mcpb_archive", lambda *a, **k: None)
+    monkeypatch.setattr(cli_release, "_check_changelog", lambda *a, **k: None)
+    monkeypatch.setattr(cli_release, "_check_formula", lambda *a, **k: None)
+
+    report = cli_release.release_check_report(repo)
+
+    assert any("definitely-not-a-cmd" in issue for issue in report.issues), report.issues
+
+
+def test_real_repo_has_no_adapter_drift() -> None:
+    """The checked-in tree must be clean, so a red run means a real regression."""
+    root = Path(__file__).resolve().parent.parent
+
+    assert adapter_issues(root) == []
